+++ src/OpenLoco/src/Windows/TerraForm.cpp
@@ -967,8 +967,8 @@ namespace OpenLoco::Ui::Windows::Terraform
                 case widx::increase_area:
                 {
                     _adjustToolSize++;
-                    if (_adjustToolSize > 10)
-                        _adjustToolSize = 10;
+                    if (_adjustToolSize > 64)
+                        _adjustToolSize = 64;
                     _clearAreaToolSize = _adjustToolSize;
                     self.invalidate();
                     break;
@@ -1061,7 +1061,10 @@ namespace OpenLoco::Ui::Windows::Terraform
 
             self.activatedWidgets |= (1ULL << widx::tool_area);
 
-            self.widgets[widx::tool_area].image = _adjustToolSize + ImageIds::tool_area;
+            if (_adjustToolSize <= 10)
+                self.widgets[widx::tool_area].image = _adjustToolSize + ImageIds::tool_area;
+            else
+                self.widgets[widx::tool_area].image = Widget::kContentNull;
 
             Widget::leftAlignTabs(self, Common::widx::tab_clear_area, Common::widx::tab_build_walls);
         }
@@ -1074,16 +1077,28 @@ namespace OpenLoco::Ui::Windows::Terraform
             self.draw(rt);
             Common::drawTabs(&self, rt);
 
+            auto& toolArea = self.widgets[widx::tool_area];
+
+            // Draw as a number if we can't fit a sprite
+            if (_adjustToolSize > 10)
+            {
+
+                auto xPos = toolArea.midX() + self.x;
+                auto yPos = toolArea.midY() + self.y - 5;
+
+                auto args = FormatArguments();
+                args.push<uint16_t>(_adjustToolSize);
+                drawingCtx.drawStringCentred(*rt, xPos, yPos, Colour::black, StringIds::tile_inspector_coord, &args);
+            }
+
             if (_raiseLandCost == 0x80000000)
                 return;
 
             if (_raiseLandCost == 0)
                 return;
 
-            auto xPos = self.widgets[widx::tool_area].left + self.widgets[widx::tool_area].right;
-            xPos /= 2;
-            xPos += self.x;
-            auto yPos = self.widgets[widx::tool_area].bottom + self.y + 5;
+            auto xPos = toolArea.midX() + self.x;
+            auto yPos = toolArea.bottom + self.y + 5;
 
             auto args = FormatArguments();
             args.push<uint32_t>(_raiseLandCost);
@@ -1114,21 +1129,24 @@ namespace OpenLoco::Ui::Windows::Terraform
             tool_area = 9,
             decrease_area,
             increase_area,
-            land_material,
-            paint_mode
+            mountain_mode,
+            paint_mode,
+            land_material
         };
 
-        const uint64_t enabledWidgets = Common::enabledWidgets | (1 << tool_area) | (1 << decrease_area) | (1 << increase_area) | (1 << land_material) | (1 << paint_mode);
+        const uint64_t enabledWidgets = Common::enabledWidgets | (1 << tool_area) | (1 << decrease_area) | (1 << increase_area) | (1 << land_material) | (1 << mountain_mode) | (1 << paint_mode);
         const uint64_t holdableWidgets = (1 << decrease_area) | (1 << increase_area);
+        static bool isMountainMode = false;
         static bool isPaintMode = false;
 
         Widget widgets[] = {
             commonWidgets(130, 105, StringIds::title_adjust_land),
-            makeWidget({ 33 + 16, 45 }, { 64, 44 }, WidgetType::wt_3, WindowColour::secondary, ImageIds::tool_area, StringIds::tooltip_adjust_land_tool),
-            makeWidget({ 34 + 16, 46 }, { 16, 16 }, WidgetType::toolbarTab, WindowColour::secondary, Gfx::recolour(ImageIds::decrease_tool_area, Colour::white), StringIds::tooltip_decrease_adjust_land_area),
-            makeWidget({ 80 + 16, 72 }, { 16, 16 }, WidgetType::toolbarTab, WindowColour::secondary, Gfx::recolour(ImageIds::increase_tool_area, Colour::white), StringIds::tooltip_increase_adjust_land_area),
-            makeWidget({ 69 + 16, 92 }, { 20, 20 }, WidgetType::wt_6, WindowColour::primary),
-            makeWidget({ 39 + 16, 92 }, { 28, 28 }, WidgetType::buttonWithImage, WindowColour::secondary, Widget::kContentNull, StringIds::tooltip_paint_landscape_tool),
+            makeWidget({ 49, 45 }, { 64, 44 }, WidgetType::wt_3, WindowColour::secondary, Widget::kContentNull, StringIds::tooltip_adjust_land_tool),
+            makeWidget({ 50, 46 }, { 16, 16 }, WidgetType::toolbarTab, WindowColour::secondary, Gfx::recolour(ImageIds::decrease_tool_area, Colour::white), StringIds::tooltip_decrease_adjust_land_area),
+            makeWidget({ 96, 72 }, { 16, 16 }, WidgetType::toolbarTab, WindowColour::secondary, Gfx::recolour(ImageIds::increase_tool_area, Colour::white), StringIds::tooltip_increase_adjust_land_area),
+            makeWidget({ 57, 92 }, { 24, 24 }, WidgetType::buttonWithImage, WindowColour::secondary, ImageIds::construction_slope_up, StringIds::mountainModeTooltip),
+            makeWidget({ 83, 92 }, { 24, 24 }, WidgetType::buttonWithImage, WindowColour::secondary, ImageIds::paintbrush, StringIds::tooltip_paint_landscape_tool),
+            makeWidget({ 112, 94 }, { 20, 20 }, WidgetType::wt_6, WindowColour::primary),
             widgetEnd(),
         };
 
@@ -1143,12 +1161,12 @@ namespace OpenLoco::Ui::Windows::Terraform
         // 0x004BBBF7
         static void tabReset(Window* self)
         {
-            Input::toolSet(
-                self,
-                isPaintMode
-                    ? static_cast<uint16_t>(widx::paint_mode)
-                    : static_cast<uint16_t>(Common::widx::panel),
-                CursorId::landTool);
+            if (isPaintMode)
+                Input::toolSet(self, widx::paint_mode, CursorId::landTool);
+            else if (isMountainMode)
+                Input::toolSet(self, widx::mountain_mode, CursorId::landTool);
+            else
+                Input::toolSet(self, Common::widx::panel, CursorId::landTool);
 
             Input::setFlag(Input::Flags::flag6);
             for (auto i = 0; i < 32; i++)
@@ -1232,8 +1250,8 @@ namespace OpenLoco::Ui::Windows::Terraform
                 case widx::decrease_area:
                 {
                     _adjustToolSize--;
-                    if (_adjustToolSize < 0)
-                        _adjustToolSize = 0;
+                    if (_adjustToolSize < 1)
+                        _adjustToolSize = 1;
                     _adjustLandToolSize = _adjustToolSize;
                     self.invalidate();
                     break;
@@ -1242,17 +1260,47 @@ namespace OpenLoco::Ui::Windows::Terraform
                 case widx::increase_area:
                 {
                     _adjustToolSize++;
-                    if (_adjustToolSize > 10)
-                        _adjustToolSize = 10;
+                    if (_adjustToolSize > 64)
+                        _adjustToolSize = 64;
                     _adjustLandToolSize = _adjustToolSize;
+                    self.invalidate();
+                    break;
+                }
+            }
+        }
+
+        static void onMouseUp(Window& self, WidgetIndex_t widgetIndex)
+        {
+            switch (widgetIndex)
+            {
+                case Common::widx::close_button:
+                    WindowManager::close(&self);
+                    break;
+
+                case Common::widx::tab_adjust_land:
+                case Common::widx::tab_adjust_water:
+                case Common::widx::tab_build_walls:
+                case Common::widx::tab_clear_area:
+                case Common::widx::tab_plant_trees:
+                    Common::switchTab(&self, widgetIndex);
+                    break;
+
+                case widx::mountain_mode:
+                {
+                    isMountainMode = !isMountainMode;
+                    isPaintMode = false;
+                    tabReset(&self);
                     self.invalidate();
                     break;
                 }
 
                 case widx::paint_mode:
                 {
+                    isMountainMode = false;
                     isPaintMode = !isPaintMode;
                     tabReset(&self);
+                    self.invalidate();
+                    break;
                 }
             }
         }
@@ -1279,7 +1327,7 @@ namespace OpenLoco::Ui::Windows::Terraform
             auto centre = (pointA + pointB) / 2;
             GameCommands::setErrorTitle(StringIds::error_cant_lower_land_here);
 
-            if (_adjustToolSize == 0)
+            if (isMountainMode)
             {
                 GameCommands::LowerRaiseLandMountainArgs args{};
                 args.centre = centre;
@@ -1311,7 +1359,7 @@ namespace OpenLoco::Ui::Windows::Terraform
             auto centre = (pointA + pointB) / 2;
             GameCommands::setErrorTitle(StringIds::error_cant_raise_land_here);
 
-            if (_adjustToolSize == 0)
+            if (isMountainMode)
             {
                 GameCommands::LowerRaiseLandMountainArgs args{};
                 args.centre = centre;
@@ -1439,6 +1487,7 @@ namespace OpenLoco::Ui::Windows::Terraform
                     onPaintToolUpdate(self, widgetIndex, x, y);
                     break;
 
+                case widx::mountain_mode:
                 case Common::widx::panel:
                     onAdjustLandToolUpdate(widgetIndex, x, y);
                     break;
@@ -1472,12 +1521,16 @@ namespace OpenLoco::Ui::Windows::Terraform
                 case widx::paint_mode:
                     paintLand();
                     break;
+
+                case widx::mountain_mode:
                 case Common::widx::panel:
+                {
                     if (!Input::hasMapSelectionFlag(Input::MapSelectionFlags::enable))
                         return;
 
                     ToolManager::setToolCursor(CursorId::upDownArrow);
                     break;
+                }
             }
         }
 
@@ -1495,6 +1548,8 @@ namespace OpenLoco::Ui::Windows::Terraform
                     }
                     break;
                 }
+
+                case widx::mountain_mode:
                 case Common::widx::panel:
                 {
                     auto window = WindowManager::findAt(x, y);
@@ -1548,13 +1603,15 @@ namespace OpenLoco::Ui::Windows::Terraform
         {
             switch (widgetIndex)
             {
+                case widx::mountain_mode:
                 case widx::paint_mode:
                 case Common::widx::panel:
+                {
                     TileManager::mapInvalidateSelectionRect();
-
                     Input::resetMapSelectionFlag(Input::MapSelectionFlags::enable);
                     ToolManager::setToolCursor(CursorId::landTool);
                     break;
+                }
             }
         }
 
@@ -1565,27 +1622,28 @@ namespace OpenLoco::Ui::Windows::Terraform
 
             self.activatedWidgets |= (1ULL << widx::tool_area);
 
+            if (isMountainMode)
+                self.activatedWidgets |= (1 << widx::mountain_mode);
+            else
+                self.activatedWidgets &= ~(1 << widx::mountain_mode);
+
+            if (isPaintMode)
+                self.activatedWidgets |= (1 << widx::paint_mode);
+            else
+                self.activatedWidgets &= ~(1 << widx::paint_mode);
+
+            auto landObj = ObjectManager::get<LandObject>(_lastSelectedLand);
+            auto pixelColour = static_cast<Colour>(Gfx::getG1Element(landObj->mapPixelImage)->offset[0]);
+            self.widgets[widx::paint_mode].image = Gfx::recolour2(ImageIds::paintbrush, Colour::white, pixelColour);
+
             if (isPaintMode)
             {
-                self.activatedWidgets |= (1 << widx::paint_mode);
+                self.widgets[widx::land_material].type = WidgetType::wt_6;
+                self.widgets[widx::land_material].image = landObj->mapPixelImage + OpenLoco::Land::ImageIds::landscape_generator_tile_icon;
             }
             else
             {
-                self.activatedWidgets &= ~(1 << widx::paint_mode);
-            }
-
-            self.widgets[widx::tool_area].image = _adjustToolSize + ImageIds::tool_area;
-
-            self.widgets[widx::land_material].type = WidgetType::none;
-
-            // CHANGE: Allows the player to select which material is used in the adjust land tool outside of editor mode.
-            if (_adjustToolSize != 0)
-            {
-                self.widgets[widx::land_material].type = WidgetType::wt_6;
-
-                auto landObj = ObjectManager::get<LandObject>(_lastSelectedLand);
-
-                self.widgets[widx::land_material].image = landObj->mapPixelImage + OpenLoco::Land::ImageIds::landscape_generator_tile_icon;
+                self.widgets[widx::land_material].type = WidgetType::none;
             }
 
             Widget::leftAlignTabs(self, Common::widx::tab_clear_area, Common::widx::tab_build_walls);
@@ -1595,19 +1653,59 @@ namespace OpenLoco::Ui::Windows::Terraform
         static void draw(Window& self, Gfx::RenderTarget* rt)
         {
             auto& drawingCtx = Gfx::getDrawingEngine().getDrawingContext();
-
-            auto skin = ObjectManager::get<InterfaceSkinObject>();
-            auto imgId = skin->img;
-            self.widgets[widx::paint_mode].image = imgId + InterfaceSkin::ImageIds::tab_colour_scheme_frame0;
-
             self.draw(rt);
 
             Common::drawTabs(&self, rt);
 
-            auto xPos = self.widgets[widx::tool_area].left + self.widgets[widx::tool_area].right;
-            xPos /= 2;
-            xPos += self.x;
-            auto yPos = self.widgets[widx::tool_area].bottom + self.y + 28;
+            auto& toolArea = self.widgets[widx::tool_area];
+
+            // Draw land tool size as a grid
+            if (_adjustToolSize < 10)
+            {
+                // For mountain mode, we first draw the background grid
+                if (isMountainMode)
+                {
+                    auto areaImage = ImageId(ImageIds::tool_area);
+                    Ui::Point placeForImage(toolArea.left + self.x, toolArea.top + self.y);
+
+                    if ((_adjustToolSize & 1) == 0)
+                    {
+                        // For even sizes, we need to draw the image twice
+                        // TODO: replace with proper grid images
+                        placeForImage -= { 4, 0 };
+                        drawingCtx.drawImage(*rt, placeForImage, areaImage);
+
+                        placeForImage += { 8, 0 };
+                        drawingCtx.drawImage(*rt, placeForImage, areaImage);
+                    }
+                    else
+                    {
+                        // For odd sizes, we just need the one
+                        drawingCtx.drawImage(*rt, placeForImage, areaImage);
+                    }
+                }
+
+                // Draw tool size
+                if (!isMountainMode || _adjustToolSize > 1)
+                {
+                    auto areaImage = ImageId(ImageIds::tool_area).withIndexOffset(_adjustToolSize);
+                    Ui::Point placeForImage(toolArea.left + self.x, toolArea.top + self.y);
+                    drawingCtx.drawImage(*rt, placeForImage, areaImage);
+                }
+            }
+            // Or draw as a number, if we can't fit a sprite
+            else
+            {
+                auto xPos = toolArea.midX() + self.x;
+                auto yPos = toolArea.midY() + self.y - 5;
+
+                auto args = FormatArguments();
+                args.push<uint16_t>(_adjustToolSize);
+                drawingCtx.drawStringCentred(*rt, xPos, yPos, Colour::black, StringIds::tile_inspector_coord, &args);
+            }
+
+            auto xPos = toolArea.midX() + self.x;
+            auto yPos = toolArea.bottom + self.y + 28;
 
             if (_raiseLandCost != 0x80000000)
             {
@@ -1635,7 +1733,7 @@ namespace OpenLoco::Ui::Windows::Terraform
         static void initEvents()
         {
             events.onClose = onClose;
-            events.onMouseUp = Common::onMouseUp;
+            events.onMouseUp = onMouseUp;
             events.onResize = onResize;
             events.onMouseDown = onMouseDown;
             events.onDropdown = onDropdown;
@@ -1711,8 +1809,8 @@ namespace OpenLoco::Ui::Windows::Terraform
                 case widx::increase_area:
                 {
                     _adjustToolSize++;
-                    if (_adjustToolSize > 10)
-                        _adjustToolSize = 10;
+                    if (_adjustToolSize > 64)
+                        _adjustToolSize = 64;
                     _adjustWaterToolSize = _adjustToolSize;
                     self.invalidate();
                     break;
@@ -1885,7 +1983,10 @@ namespace OpenLoco::Ui::Windows::Terraform
 
             self.activatedWidgets |= (1ULL << widx::tool_area);
 
-            self.widgets[widx::tool_area].image = _adjustToolSize + ImageIds::tool_area;
+            if (_adjustToolSize <= 10)
+                self.widgets[widx::tool_area].image = _adjustToolSize + ImageIds::tool_area;
+            else
+                self.widgets[widx::tool_area].image = Widget::kContentNull;
 
             Widget::leftAlignTabs(self, Common::widx::tab_clear_area, Common::widx::tab_build_walls);
         }
@@ -1898,10 +1999,21 @@ namespace OpenLoco::Ui::Windows::Terraform
             self.draw(rt);
             Common::drawTabs(&self, rt);
 
-            auto xPos = self.widgets[widx::tool_area].left + self.widgets[widx::tool_area].right;
-            xPos /= 2;
-            xPos += self.x;
-            auto yPos = self.widgets[widx::tool_area].bottom + self.y + 5;
+            auto& toolArea = self.widgets[widx::tool_area];
+
+            // Draw as a number if we can't fit a sprite
+            if (_adjustToolSize > 10)
+            {
+                auto xPos = toolArea.midX() + self.x;
+                auto yPos = toolArea.midY() + self.y - 5;
+
+                auto args = FormatArguments();
+                args.push<uint16_t>(_adjustToolSize);
+                drawingCtx.drawStringCentred(*rt, xPos, yPos, Colour::black, StringIds::tile_inspector_coord, &args);
+            }
+
+            auto xPos = toolArea.midX() + self.x;
+            auto yPos = toolArea.bottom + self.y + 5;
 
             if (_raiseWaterCost != 0x80000000)
             {